refactor: remove language fetch outside of app
